Max sum subarray by removing at most one element
You are given array A of size n. You need to find the maximum-sum sub-array with the condition that you are allowed to skip at most one element.
Input Format:
The first line of input contains an integer T denoting the number of test cases. The description of T test cases follows. The first line of each test case contains a single integer n denoting the size of the array. The second line contains n space-separated integers A1, A2, ..., An denoting the elements of the array.
Output Format:
For each test case, in a new line, print the maximum sum.
Your Task:
This is a function problem. The driver code takes input. Your task is to complete the function maxSumSubarray that take array and size as parameters and returns the maximum sum.
Expected Time Complexity: O(N).
Expected Auxiliary Space: O(N).

Constraints:
1 <= T <= 100
1 <= n <= 100
-103 <= Ai<= 103
Example:
Input:
2
5
1 2 3 -4 5
8
-2 -3 4 -1 -2 1 5 -3
Output:
11
9